UPSTREAM: drm/vkms: Update pixels accessor to support packed and multi-plane formats.
Introduce the usage of block_h/block_w to compute the offset and the
pointer of a pixel. The previous implementation was specialized for
planes with block_h == block_w == 1. To avoid confusion and allow easier
implementation of tiled formats. It also remove the usage of the
deprecated format field `cpp`.

Introduce the plane_index parameter to get an offset/pointer on a
different plane.

Bug: 382271731
Test: Set vkms as drm display driver and
      verified no errors in logcat and dmesg

Acked-by: Maíra Canal <mairacanal@riseup.net>
Reviewed-by: José Expósito <jose.exposito89@gmail.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20241118-yuv-v14-5-2dbc2f1e222c@bootlin.com
(cherry picked from commit d6dc39e495678d54a8125e1ae08b3ca35c306426)
Change-Id: I2988edecd9e302e927175790825e4f55d45f9f15
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: Paz Zcharya <pazz@google.com>