UPSTREAM: drm/vkms: Introduce pixel_read_direction enum
The pixel_read_direction enum is useful to describe the reading direction
in a plane. It avoids using the rotation property of DRM, which not
practical to know the direction of reading.
This patch also introduce two helpers, one to compute the
pixel_read_direction from the DRM rotation property, and one to compute
the step, in byte, between two successive pixel in a specific direction.

Bug: 382271731
Test: Set vkms as drm display driver and
      verified no errors in logcat and dmesg

Acked-by: Maíra Canal <mairacanal@riseup.net>
Reviewed-by: José Expósito <jose.exposito89@gmail.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20241118-yuv-v14-7-2dbc2f1e222c@bootlin.com
(cherry picked from commit b52fd27356af947f80b0c0855691e42a4c02ee79)
Change-Id: I4e2c77b90334d7bd0b2b56f5f1473217c7f10e28
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: Paz Zcharya <pazz@google.com>