FROMGIT: drm/vkms: Allow to configure multiple CRTCs
Add a list of CRTCs to vkms_config and helper functions to add and
remove as many CRTCs as wanted.

For backwards compatibility, add one CRTC to the default configuration.

A future patch will allow to attach planes and CRTCs, but for the
moment there are no changes in the way the output is configured.

Reviewed-by: Louis Chauvet <louis.chauvet@bootlin.com>
Co-developed-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: José Expósito <jose.exposito89@gmail.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20250218101214.5790-10-jose.exposito89@gmail.com
Signed-off-by: Maxime Ripard <mripard@kernel.org>
(cherry picked from commit 600df32dac40ad4b1069615a970063f0b15abfda
 https://anongit.freedesktop.org/git/drm/drm-misc.git drm-misc-next)

Bug: 414366730
Test: atest VtsHalGraphicsComposer3_TargetTest

Change-Id: I18b4053360bc2b99c8c5da77b9a7284db002343d
Signed-off-by: Paz Zcharya <pazz@google.com>