FROMGIT: drm/vkms: Allow to configure multiple encoders
Add a list of encoders to vkms_config and helper functions to add and
remove as many encoders as wanted.

For backwards compatibility, add one encoder to the default
configuration.

A future patch will allow to attach encoders and CRTCs, but for the
moment there are no changes in the way the output is configured.

Reviewed-by: Louis Chauvet <louis.chauvet@bootlin.com>
Co-developed-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: José Expósito <jose.exposito89@gmail.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20250218101214.5790-12-jose.exposito89@gmail.com
Signed-off-by: Maxime Ripard <mripard@kernel.org>
(cherry picked from commit f60a183dc9105e3dc5120d5aa03b294d182965fb
 https://anongit.freedesktop.org/git/drm/drm-misc.git drm-misc-next)

Bug: 414366730
Test: atest VtsHalGraphicsComposer3_TargetTest

Change-Id: Iaf39fc4c544c9d0dc0fa9650656ee4b74d933d82
Signed-off-by: Paz Zcharya <pazz@google.com>