FROMGIT: drm/vkms: Allow to configure multiple connectors
Add a list of connectors to vkms_config and helper functions to add and
remove as many connectors as wanted.

For backwards compatibility, add one enabled connector to the default
configuration.

A future patch will allow to attach connectors and encoders, but for the
moment there are no changes in the way the output is configured.

Reviewed-by: Louis Chauvet <louis.chauvet@bootlin.com>
Co-developed-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: Louis Chauvet <louis.chauvet@bootlin.com>
Signed-off-by: José Expósito <jose.exposito89@gmail.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20250218101214.5790-14-jose.exposito89@gmail.com
Signed-off-by: Maxime Ripard <mripard@kernel.org>
(cherry picked from commit da38c72018e28b08197975cec7ffd5e07f189a9c
 https://anongit.freedesktop.org/git/drm/drm-misc.git drm-misc-next)

Bug: 414366730
Test: atest VtsHalGraphicsComposer3_TargetTest

Change-Id: I1e33a1a1da3a9400193030354f8f2e7d73a827ac
Signed-off-by: Paz Zcharya <pazz@google.com>