ANDROID: virtio_dxgkrnl: throttle killed process
When a process is killed, it will close the /dev/dxg device and the
virtio-dxgkrnl driver will destroy allocations and resources
associated with that process. It will do this via the normal
dxgvmb_send_* functions, which for the most part are synchronous waiting
for the device to send a response.

However, these wait via completions and we wait on those completions by
calling wait_for_completion_interruptible, and if the process is being
killed this function will return immediately. If the process has a lot
of allocations this could quickly fill up the virtqueue and result in
dropped commands.

The dxgvmb_send_sync_msg function will now check if
wait_for_completion_interruptible returns -ERESTARTSYS, and if so it
will do an uninterruptible wait with a timeout of 500 usecs. If the
device is keeping up with the virtqueue then this should return quickly,
otherwise it will return in 500 usecs and continue.

I have tested this by creating an application that allocates 4k
allocations and then sleeps. I then kill that application. Before, this
would cause the virtqueue to be full and commands to be dropped, but now
the killed application waits periodically while it's destroying the
allocations.

Bug: b/415109094
Test: killed process with lots of allocations
Signed-off-by: Colin Downs-Razouk <colindr@google.com>
Change-Id: Ic5961231c6254494e5c4939b1c8d96aec919b6cc